hide symbols also on first build
The option to build shared libs was defined after it was checked for the first
time, so on the first run it would be undefined and thus evaluating to false.
The symbols were therefore not hidden. On a subsequent run of CMake (e.g. by
changing any other option or touching CMakeLists.txt) the variable is set from
the cache and the visibility option was switched to hidden.